requirements.txt
matplotlib==3.3.2
pandas==1.1.2
python-dotenv==0.14.0
requests==2.24.0
SQLAlchemy==1.3.19
wordcloud==1.8.0
